[Gold III] Title: 나머지 합, Time: 124 ms, Memory: 9836 KB -BaekjoonHub

+++ 백준/Gold/10986. 나머지 합/나머지 합.cc
@@ -6,7 +6,7 @@ const int MMX = 1000;
 int N, M;
 int arr[NMX];
 int P[NMX + 1];
-long long cnt[MMX];
+int cnt[MMX];
 
 int main(int argc, char const *argv[])
 {
@@ -48,9 +48,9 @@ int main(int argc, char const *argv[])
     long long ans = 0;
     for (int i = 0; i < M; i++)
     {
-        long long n = cnt[i];
+        int n = cnt[i];
 
-        ans += n * (n - 1) / 2;
+        ans += (long long)n * (n - 1) / 2;
     }
 
     cout << ans;
